OrderState changes
- Notes on relays now include payment and order status publicly for
  aggregation

- Invoicer holds a more complete list of live orders for bettter
  handling

- Removed parallel cahnnels of invoice and improved concurrency
