+++ invoicer/src/state.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use anyhow::anyhow;
 use fuente::models::{
-    AdminConfiguration, AdminConfigurationType, AdminServerRequest, CommerceProfile, DriverProfile,
+    AdminConfiguration, AdminConfigurationType, AdminServerRequest, CommerceProfile,
     OrderInvoiceState, ProductMenu, TEST_PUB_KEY,
 };
 use nostro2::{
@@ -148,7 +148,7 @@ impl InvoicerStateLock {
     }
     pub async fn add_live_order(&self, order: OrderInvoiceState) -> anyhow::Result<()> {
         let mut orders = self.lock().await;
-        orders.live_orders.new_order(order.id(), order)?;
+        orders.live_orders.new_order(order.order_id(), order)?;
         Ok(())
     }
     pub async fn find_live_order(&self, order_id: &str) -> Option<OrderInvoiceState> {
@@ -158,41 +158,10 @@ impl InvoicerStateLock {
         let mut orders = self.lock().await;
         let order = orders
             .live_orders
-            .get_mut_order(&order_update.id())
+            .get_mut_order(&order_update.order_id())
             .ok_or(anyhow!("Order not found"))?;
         *order = order.to_owned();
         Ok(())
-    }
-    pub async fn handle_courier_updates(
-        &self,
-        inner_note: NostrNote,
-        outer_note: NostrNote,
-    ) -> anyhow::Result<OrderInvoiceState> {
-        let bot_state = self.lock_owned().await;
-        bot_state
-            .admin_config
-            .check_couriers_whitelist(&inner_note.pubkey)?;
-        let updated_order: OrderInvoiceState =
-            OrderInvoiceState::try_from(inner_note.content.to_string())?;
-        // Check if order is part of live orders
-        let mut live_order = bot_state
-            .live_orders
-            .get_order(&updated_order.id())
-            .ok_or(anyhow!("Order not found"))?;
-        // Check if the courier is already assigned
-        if live_order.get_courier().is_none() {
-            let new_courier = updated_order
-                .get_courier()
-                .ok_or(anyhow!("No courier found"))?;
-            live_order.update_courier(new_courier);
-            return Ok(live_order.to_owned());
-        }
-        // Check if the update s coing from asigned courier
-        if outer_note.pubkey == live_order.get_courier().unwrap().pubkey {
-            info!("Order updated by courier");
-            return Ok(updated_order.to_owned());
-        }
-        Err(anyhow!("Invalid courier update"))
     }
     pub async fn sign_updated_config(
         &self,
